added makefile and performed lint cleaning operations. This should've been a separate issue, but since I am the only contributor :'-( , it wouldnt hurt. close #4
