+++ auth/auth.go
@@ -3,7 +3,7 @@ package auth
 import (
 	"encoding/json"
 	"fmt"
-	"github.com/martinomburajr/gogist/config"
+	"github.com/martinomburajr/gist/config"
 	"io/ioutil"
 	"log"
 	"net/http"
@@ -11,11 +11,25 @@ import (
 )
 
 var (
+	//RedirectURI is the URI the Github OAuth flow redirects to
 	RedirectURI = fmt.Sprintf("http://localhost:%d/auth/github/callback", config.PORT)
+
+	//BaseURL is the base URL to perform a login, this URL does not point to anything by itself,
+	// it needs to be composed with other information. See AuthURL for the full Login URL
 	BaseURL = "https://github.com/login/oauth/authorize"
+
+	//ClientID represents the client id - this should never be placed in code but rather injected via a variable
 	ClientID = ""
+
+	//ClientSecret represents the application client secret - THIS SHOULD NEVER BE PLACED IN CODE BUT INJECTED VIA
+	// ENVIRONMENT VARIABLE
 	ClientSecret = ""
+
+	//AuthURL is the final URL used to perform a login
 	AuthURL = fmt.Sprintf("%s?client_id=%s&redirect_uri=%s", BaseURL, ClientID, RedirectURI)
+
+	//Session is a singleton variable that holds all authentication and config based information for a session to
+	// succeed.
 	Session = SessionObj{}
 )
 
@@ -44,7 +58,7 @@ func RedirectHandler(w http.ResponseWriter, r *http.Request) {
 	log.Print("I AM HERE REDIRECTED")
 	err := r.ParseForm()
 	if err != nil {
-		fmt.Fprint(os.Stdout, "could not parse query: %v", err)
+		fmt.Fprintf(os.Stdout, "could not parse query: %s", err.Error())
 		w.WriteHeader(http.StatusBadRequest)
 	}
 	code := r.FormValue("code")
@@ -52,7 +66,7 @@ func RedirectHandler(w http.ResponseWriter, r *http.Request) {
 	reqURL := fmt.Sprintf("https://github.com/login/oauth/access_token?client_id=%s&client_secret=%s&code=%s", ClientID, ClientSecret, code)
 	req, err := http.NewRequest(http.MethodPost, reqURL, nil)
 	if err != nil {
-		fmt.Fprintf(os.Stdout, "could not retrieve http request: %v", err)
+		fmt.Fprintf(os.Stdout, "could not retrieve http request: %s", err.Error())
 		w.WriteHeader(http.StatusBadRequest)
 	}
 
@@ -61,7 +75,7 @@ func RedirectHandler(w http.ResponseWriter, r *http.Request) {
 
 	res, err := http.DefaultClient.Do(req)
 	if err != nil {
-		fmt.Fprintf(os.Stdout, "could not send HTTP request: %v", err)
+		fmt.Fprintf(os.Stdout, "could not send HTTP request: %s", err.Error())
 		w.WriteHeader(http.StatusInternalServerError)
 	}
 
@@ -69,7 +83,7 @@ func RedirectHandler(w http.ResponseWriter, r *http.Request) {
 	// Parse the request body into the `OAuthAccessResponse` struct
 	var t OAuthAccessResponse
 	if err := json.NewDecoder(res.Body).Decode(&t); err != nil {
-		fmt.Fprintf(os.Stdout, "could not parse JSON response: %v", err)
+		fmt.Fprintf(os.Stdout, "could not parse JSON response: %s", err.Error())
 		w.WriteHeader(http.StatusBadRequest)
 	}
 	Session.AccessToken = t.AccessToken
@@ -78,10 +92,13 @@ func RedirectHandler(w http.ResponseWriter, r *http.Request) {
 	w.Write([]byte("OK"))
 }
 
+//SessionObj is a type that contains session based information for authentication based actions to work
 type SessionObj struct {
 	AccessToken string `json:"access_token"`
 	Client *http.Client
 }
+
+//OAuthAccessResponse embodies a response from the GitHub OAuth server with the AccessToken if authorized.
 type OAuthAccessResponse struct {
 	AccessToken string `json:"access_token"`
 }
