+++ config/config.go
@@ -1,5 +1,6 @@
 package config
 
 const (
+	//PORT is the port at which the server can be accessed
 	PORT = 8089
 )
