+++ cover/auth.cov
@@ -1,0 +1,1 @@
+mode: count
