+++ cover/coverage.cov
@@ -1,0 +1,58 @@
+github.com/martinomburajr/gist/handlers.go:19.65,21.2 0 0
+github.com/martinomburajr/gist/handlers.go:23.93,26.27 3 0
+github.com/martinomburajr/gist/handlers.go:37.2,37.30 1 0
+github.com/martinomburajr/gist/handlers.go:26.27,27.33 1 0
+github.com/martinomburajr/gist/handlers.go:27.33,29.18 2 0
+github.com/martinomburajr/gist/handlers.go:29.18,32.5 2 0
+github.com/martinomburajr/gist/handlers.go:32.9,34.5 1 0
+github.com/martinomburajr/gist/handlers.go:40.54,43.30 3 0
+github.com/martinomburajr/gist/handlers.go:54.2,54.18 1 0
+github.com/martinomburajr/gist/handlers.go:43.30,48.17 3 0
+github.com/martinomburajr/gist/handlers.go:48.18,50.4 0 0
+github.com/martinomburajr/gist/handlers.go:50.8,52.4 1 0
+github.com/martinomburajr/gist/handlers.go:57.44,59.81 2 0
+github.com/martinomburajr/gist/handlers.go:63.2,63.16 1 0
+github.com/martinomburajr/gist/handlers.go:66.2,66.29 1 0
+github.com/martinomburajr/gist/handlers.go:69.2,69.14 1 0
+github.com/martinomburajr/gist/handlers.go:59.81,62.3 2 0
+github.com/martinomburajr/gist/handlers.go:63.16,64.13 1 0
+github.com/martinomburajr/gist/handlers.go:66.29,68.3 1 0
+github.com/martinomburajr/gist/main.go:13.13,31.2 5 0
+github.com/martinomburajr/gist/main.go:33.97,34.54 1 0
+github.com/martinomburajr/gist/main.go:34.54,39.18 1 0
+github.com/martinomburajr/gist/main.go:39.18,41.4 1 0
+github.com/martinomburajr/gist/main.go:47.58,51.16 3 0
+github.com/martinomburajr/gist/main.go:51.16,53.3 1 0
+github.com/martinomburajr/gist/main.go:57.20,59.2 0 0
+github.com/martinomburajr/gist/handlers.go:23.93,26.27 3 0
+github.com/martinomburajr/gist/handlers.go:37.2,37.30 1 0
+github.com/martinomburajr/gist/handlers.go:26.27,27.33 1 0
+github.com/martinomburajr/gist/handlers.go:27.33,29.18 2 0
+github.com/martinomburajr/gist/handlers.go:29.18,32.5 2 0
+github.com/martinomburajr/gist/handlers.go:32.9,34.5 1 0
+github.com/martinomburajr/gist/handlers.go:40.54,43.30 3 0
+github.com/martinomburajr/gist/handlers.go:54.2,54.18 1 0
+github.com/martinomburajr/gist/handlers.go:43.30,48.17 3 0
+github.com/martinomburajr/gist/handlers.go:48.18,50.4 0 0
+github.com/martinomburajr/gist/handlers.go:50.8,52.4 1 0
+github.com/martinomburajr/gist/handlers.go:57.44,59.81 2 0
+github.com/martinomburajr/gist/handlers.go:63.2,63.16 1 0
+github.com/martinomburajr/gist/handlers.go:66.2,66.29 1 0
+github.com/martinomburajr/gist/handlers.go:69.2,69.14 1 0
+github.com/martinomburajr/gist/handlers.go:59.81,62.3 2 0
+github.com/martinomburajr/gist/handlers.go:63.16,64.13 1 0
+github.com/martinomburajr/gist/handlers.go:66.29,68.3 1 0
+github.com/martinomburajr/gist/main.go:13.13,31.2 5 0
+github.com/martinomburajr/gist/main.go:33.97,34.54 1 0
+github.com/martinomburajr/gist/main.go:34.54,39.18 1 0
+github.com/martinomburajr/gist/main.go:39.18,41.4 1 0
+github.com/martinomburajr/gist/main.go:47.58,51.16 3 0
+github.com/martinomburajr/gist/main.go:51.16,53.3 1 0
+github.com/martinomburajr/gist/main.go:57.20,59.2 0 0
+github.com/martinomburajr/gist/handlers.go:9.65,11.2 0 0
+github.com/martinomburajr/gist/main.go:13.13,32.2 5 0
+github.com/martinomburajr/gist/main.go:37.97,38.54 1 0
+github.com/martinomburajr/gist/main.go:38.54,43.18 1 0
+github.com/martinomburajr/gist/main.go:43.18,45.4 1 0
+github.com/martinomburajr/gist/main.go:52.58,56.16 3 0
+github.com/martinomburajr/gist/main.go:56.16,58.3 1 0
