+++ cover/gist.cov
@@ -1,0 +1,8 @@
+mode: count
+github.com/martinomburajr/gist/handlers.go:9.65,11.2 0 0
+github.com/martinomburajr/gist/main.go:13.13,32.2 5 0
+github.com/martinomburajr/gist/main.go:37.97,38.54 1 0
+github.com/martinomburajr/gist/main.go:38.54,43.18 1 0
+github.com/martinomburajr/gist/main.go:43.18,45.4 1 0
+github.com/martinomburajr/gist/main.go:52.58,56.16 3 0
+github.com/martinomburajr/gist/main.go:56.16,58.3 1 0
