+++ gists/gistparser.go
@@ -8,12 +8,16 @@ import (
 	"strings"
 )
 
-
+//GistParser is an object that represents the items required to parse a gist. Typically the filepath and its contents
 type GistParser struct {
-	Filepath string `json:"badfilepath"`
+	Filepath string `json:"filepath"`
 	fileContents []byte
 }
 
+//ToGist is an accumulator method that performs multiple sub functions.
+// It is responsible for the creation of a gist file, this involves, parsing the  file to determine if it s gistable,
+// if it is gistable, relavant metadata is extracted from the file, the file is then converted to the GistFile type.
+// This is the final form before other operations can be performed on it once it is a GistFile.
 func (g *GistParser) ToGist() (*GistFile, error) {
 	err := g.IsGistable()
 	if err != nil {
@@ -48,6 +52,7 @@ func (g *GistParser) ToGist() (*GistFile, error) {
 	}, nil
 }
 
+//GetFileBody extracts the body of a gist from the file.
 func (g *GistParser) GetFileBody() (*GistFileBody, error) {
 	if len(g.fileContents) == 0 {
 		err := g.Reader()
@@ -61,7 +66,7 @@ func (g *GistParser) GetFileBody() (*GistFileBody, error) {
 
 // IsGistable checks to a see a certain file conforms to the GOGIST standard.
 // If the file does not contain the "GOGIST" label
-// in a comments section in the file. It is deemed ungistable meaning, gogist will not create a gist for the user in that regard.
+// in a comments section in the file. It is deemed ungistable meaning, gist will not create a gist for the user in that regard.
 // If no error is presented, one can assume that it is a gistable file.
 func (g *GistParser) IsGistable() error {
 	err := g.Reader()
@@ -71,18 +76,18 @@ func (g *GistParser) IsGistable() error {
 
 	containsStart := bytes.Contains(bytes.ToLower(g.fileContents), bytes.ToLower([]byte("start GOGIST")))
 	if !containsStart {
-		return fmt.Errorf("is not a suitable GOGIST file. Add the following string 'start GOGIST' inside a comment section at the top of the file to mark it as a file gogist can gist ;-)")
+		return fmt.Errorf("is not a suitable GOGIST file. Add the following string 'start GOGIST' inside a comment section at the top of the file to mark it as a file gist can gist ;-)")
 	}
 
 	containsEnd := bytes.Contains(bytes.ToLower(g.fileContents), bytes.ToLower([]byte("end GOGIST")))
 	if !containsEnd {
-		return fmt.Errorf("is not a suitable GOGIST file. Add the following string 'end GOGIST' inside a comment section at the top of the file to mark it as a file gogist can gist ;-). This should be after the start Gogist section")
+		return fmt.Errorf("is not a suitable GOGIST file. Add the following string 'end GOGIST' inside a comment section at the top of the file to mark it as a file gist can gist ;-). This should be after the start Gogist section")
 	}
 
 	return nil
 }
 
-// getAuthor returns the Author information. The author must precede the GOGIST label, and must contain all runes within the word
+// GetAuthor returns the Author information. The author must precede the GOGIST label, and must contain all runes within the word
 // "AUTHOR". This is CASE sensitive
 //
 // This is the format shown below. Email is optional. Anything after newline carriage return is considered not part of the author label
@@ -91,7 +96,7 @@ func (g *GistParser) IsGistable() error {
 //	Author: I am some author <hereismy@email.com>
 //  Description: Some awesome gist
 //  Public: true
-//  end gogist
+//  end gist
 //	*/
 //  returns I am some author <hereismy@email.com>
 //
@@ -103,7 +108,8 @@ func (g *GistParser) GetAuthor() (string, error) {
 	return g.getContent(lines, "author")
 }
 
-// getDescription returns the Author information. The Description must precede the GOGIST label, and must contain all runes within the word
+// GetDescription returns the Author information. The Description must precede the GOGIST label,
+// and must contain all runes within the word
 // "Description". This is CASE sensitive
 //
 // This is the format shown below. Email is optional. Anything after newline carriage return is considered not part of the Description label so ensure description is all in a single line.
@@ -112,7 +118,7 @@ func (g *GistParser) GetAuthor() (string, error) {
 //	Author: I am some author <hereismy@email.com>
 //  Description: Some awesome gist
 //  Public: true
-//  end gogist
+//  end gist
 //	*/
 //  returns Some awesome gist
 //
@@ -124,7 +130,8 @@ func (g *GistParser) GetDescription() (string, error) {
 	return g.getContent(lines, "description")
 }
 
-// getPublic returns the isPublic information. The Public must precede the GOGIST label, and must contain all runes within the word
+// GetPublic returns the isPublic information. The Public must precede the GOGIST label,
+// and must contain all runes within the word
 // "AUTHOR". This is CASE insensitive
 //
 // This is the format shown below. Email is optional. Public is a boolean variable that can either be true or false. Its default value is true
@@ -133,7 +140,7 @@ func (g *GistParser) GetDescription() (string, error) {
 //	Author: I am some author <hereismy@email.com>
 //  Description: Some awesome gist
 //  Public: true
-//  end gogist
+//  end gist
 //	*/
 // returns true
 //
@@ -153,11 +160,11 @@ func (g *GistParser) GetPublic() (bool, error) {
 	return b, nil
 }
 
-//getGogistLines returns the lines encapsulated by the 'start gogist' and the'end gogist' labels. This is where all the important gogist metadata is found.
+//getGogistLines returns the lines encapsulated by the 'start gist' and the'end gist' labels. This is where all the important gist metadata is found.
 func (g *GistParser) getGogistLines() ([]string, error) {
 	err := g.IsGistable()
 	if err != nil {
-		return nil, fmt.Errorf("could not determine if file has 'start gogist' and the'end gogist' labels -> %s", err.Error())
+		return nil, fmt.Errorf("could not determine if file has 'start gist' and the'end gist' labels -> %s", err.Error())
 	}
 
 	buffer := bytes.NewBuffer(g.fileContents)
@@ -165,12 +172,12 @@ func (g *GistParser) getGogistLines() ([]string, error) {
 
 	startIndex := -1
 	endIndex := -1
-	for i, _ := range documentContents {
+	for i := range documentContents {
 		documentContents[i] = strings.Trim(documentContents[i], " \r\n")
-		if strings.Contains(  strings.ToLower(documentContents[i]), strings.ToLower("start gogist"))  {
+		if strings.Contains(  strings.ToLower(documentContents[i]), strings.ToLower("start gist"))  {
 			startIndex = i
 		}
-		if strings.Contains(strings.ToLower(documentContents[i]), strings.ToLower("end gogist"))  {
+		if strings.Contains(strings.ToLower(documentContents[i]), strings.ToLower("end gist"))  {
 			endIndex = i
 			break
 		}
@@ -179,7 +186,7 @@ func (g *GistParser) getGogistLines() ([]string, error) {
 	return documentContents[startIndex:endIndex+1], nil
 }
 
-//getContent takes in the gogist section obtained after running getGogistLines, and obtaining the exact metadata section. key represents a  key in a key-value pair. e.g. author or description are valid keys
+//getContent takes in the gist section obtained after running getGogistLines, and obtaining the exact metadata section. key represents a  key in a key-value pair. e.g. author or description are valid keys
 func (g *GistParser) getContent(s []string, key string) (string, error) {
 	var location = -1
 	for i, v := range s {
@@ -195,9 +202,8 @@ func (g *GistParser) getContent(s []string, key string) (string, error) {
 		finstring := totalString[(len(totalString) - lenreplaced):]
 		trim := strings.Trim(finstring, " ")
 		return trim, nil
-	}  else {
-		return "", fmt.Errorf(key + " does not exist")
 	}
+	return "", fmt.Errorf(key + " does not exist")
 }
 
 
@@ -213,7 +219,7 @@ func (g *GistParser) getContent(s []string, key string) (string, error) {
 //	return len(data), nil
 //}
 
-//Reads the entire file contents using ioutil. It then appends the data to the p variable for later use.
+//Reader Reads the entire file contents using ioutil. It then appends the data to the p variable for later use.
 func (g *GistParser) Reader()  (err error) {
 	data, err := ioutil.ReadFile(g.Filepath)
 	if err != nil {
